+++ src/main/java/com/google/codeu/data/Datastore.java
@@ -45,7 +45,6 @@ public class Datastore {
     messageEntity.setProperty("text", message.getText());
     messageEntity.setProperty("timestamp", message.getTimestamp());
     messageEntity.setProperty("recipient", message.getRecipient());
-    messageEntity.setProperty("score", message.getScore());
 
     datastore.put(messageEntity);
   }
@@ -76,7 +75,7 @@ public class Datastore {
 
         float sentimentScore = entity.getProperty("score") == null? (float) 0.0 : ((Double) entity.getProperty("score")).floatValue();
 
-        Message message = new Message(id, user, text, timestamp, recipient, sentimentScore);
+        Message message = new Message(id, user, text, timestamp, recipient);
 
         messages.add(message);
       } catch (Exception e) {
@@ -111,9 +110,8 @@ public class Datastore {
         String text = (String) entity.getProperty("text");
         long timestamp = (long) entity.getProperty("timestamp");
         String recipient = (String) entity.getProperty("recipient");
-        float sentimentScore = entity.getProperty("score") == null? (float) 0.0 : ((Double) entity.getProperty("score")).floatValue();
 
-        Message message = new Message(id, user, text, timestamp, recipient, sentimentScore);
+        Message message = new Message(id, user, text, timestamp, recipient);
         messages.add(message);
       } catch (Exception e) {
         System.err.println("Error reading message.");
