+++ src/main/java/com/google/codeu/data/Message.java
@@ -27,23 +27,21 @@ public class Message {
   private String text;
   private long timestamp;
   private String recipient;
-  private float score;
 
   /**
    * Constructs a new {@link Message} posted by {@code user} with {@code text} content. Generates a
    * random ID and uses the current system time for the creation time.
    */
-  public Message(String user, String text, String recipient, float score) {
-    this(UUID.randomUUID(), user, text, System.currentTimeMillis(), recipient, score);
+  public Message(String user, String text, String recipient) {
+    this(UUID.randomUUID(), user, text, System.currentTimeMillis(), recipient);
   }
 
-  public Message(UUID id, String user, String text, long timestamp, String recipient, float score) {
+  public Message(UUID id, String user, String text, long timestamp, String recipient) {
     this.id = id;
     this.user = user;
     this.text = text;
     this.timestamp = timestamp;
     this.recipient = recipient;
-    this.score = score;
   }
 
   public UUID getId() {
@@ -70,7 +68,4 @@ public class Message {
     this.text = translatedText;
   }
 
-  public float getScore() {
-        return score;
-    }
 }
