+++ src/main/webapp/groupchat.html
@@ -38,7 +38,7 @@ limitations under the License.
     <script src="/js/navigation-loader.js"></script>
 </head>
 
-<body onload="addLoginOrLogoutLinkToNavigation();">
+<body onload="buildUI();">
     <!-- Navigation menu component -->
     <nav>
         <!-- Bootstrap nav menu template -->
@@ -60,6 +60,7 @@ limitations under the License.
         </ul>
     </nav>
     <!-- End of navigation menu component -->
+
