+++ src/main/webapp/login.html
@@ -44,7 +44,7 @@ limitations under the License.
                 <div class="card card-signin my-5">
                     <div class="card-body">
                         <h5 class="card-title text-center">Sign In</h5>
-                        <form class="form-signin" action="login" method=post>
+                        <form class="form-signin" action="/login" method=post target="_self">
                             <div class="form-label-group">
                                 <input type="email" id="inputEmail" name="inputEmail"class="form-control" placeholder="Email address"
                                     required autofocus>
